Added Missing RTG Pellet Recipes/Decay
I don't understand why the Berkelium RTG Pellet wasn't given any sort of decay to begin with, but whatever. Now it's simply a longer-lasting Americium RTG Pellet. Ironically, it even decays into Americium. I wouldn't question it that much, I actually looked it up and just went for it. You can change it later or somethin, idk.
I'm not gonna even question why it even uses Nickel instead of Iron to craft it, but I'm assuming they're supposed to be crafted later so I gave the Californium RTG Pellets the same treatment, which were somehow not craftable to begin with...
I don't know what the end goal is for these things, but at least they're obtainable and usable in the meantime...